implをmodへ移動、直交基底関数にDocsを追加
